search bar works and placed in module
