+++ javascripts/main.js
@@ -1,13 +1,14 @@
 "use strict";
 
+const highlight = require('./SearchHighlightDOM');
 const factory = require('./factory');
 const formatter = require('./formatter');
 const parkInfo = require('./parkInfo');
 const currentEvents = require('./currentEvents');
 const areaToDom = require('./areaToDom');
 const attractionToDom = require('./attractionToDom');
+const _ = require('lodash');
 // const attractionTemplate = require('../templates/areaAttraction.hbs');
-
 const $ = require('jquery');
 
 factory.getParkInfo();
@@ -15,6 +16,7 @@ factory.getAreas();
 factory.getAttractions();
 areaToDom.outputArea();
 attractionToDom.attractionByArea();
+highlight.highlightArea();
 
 // Attempting to format data
 // formatter.getData().then(data => )
@@ -93,50 +95,6 @@ setCurrentTime();
 
 
 
-
-// Search function to match user input with JSON
-$(document).ready(function(){
-    $("#search").keydown(function(e){
-        if(e.which == 13) {
-            var userSearch = $("#search").val().toLowerCase();
-            $('#search').val(''); //CLEAR INPUT
-            $(".highlight").removeClass("highlight");
-            factory.getAttractions().then( (attrData) => {
-                var matchingAttractions = [];
-                for (let i = 0; i < attrData.length; i++){
-                    var attractionNames = attrData[i].name.toLowerCase();
-                    if(attractionNames.includes(userSearch)){
-                        matchingAttractions.push(attrData[i]);
-                    }
-                }
-                //loop over matching attractions array and find the area_id key
-                getAreaID(matchingAttractions);
-                return matchingAttractions;
-            });
-        }
-    });
-});
-
-//loop over matching attractions array and find the area_id key
-var _ = require('lodash');
-
-
-function getAreaID(attractionArr) {
-    var idArr = [];
-    attractionArr.forEach(function(e){
-    idArr.push(e.area_id);
-// use area_id to give a class to area section, or to toggle highlight class
-    let newIdArr = _.uniqBy(idArr);
-    newIdArr.forEach( (area_id) =>{
-        $(`#${area_id}`)
-        .addClass("highlight");
-    });
-
-
-    console.log(newIdArr); 
-    return idArr;
-});
-}
